clinical_research_api/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
drf-spectacular==0.26.0
psycopg2-binary==2.9.6
python-decouple==3.8
django-cors-headers==4.0.0